+++ packages/validators/src/raw/__tests__/requiredIf.spec.js
@@ -1,38 +1,22 @@
 import requiredIf from '../requiredIf'
 import { T, F } from '../../../tests/fixtures'
-
-const promiseT = () => Promise.resolve(true)
-const promiseF = () => Promise.resolve(false)
+import { ref } from 'vue-demi'
 
 describe('requiredIf validator', () => {
   it('should not validate empty string when functional condition is met', () => {
-    expect(requiredIf(T)('')).resolves.toBe(false)
+    expect(requiredIf(T)('')).toBe(false)
   })
 
   it('should validate empty string when functional condition not met', () => {
-    expect(requiredIf(F)('')).resolves.toBe(true)
+    expect(requiredIf(F)('')).toBe(true)
   })
 
   it('should not validate empty string when simple boolean condition is met', () => {
-    expect(requiredIf('prop')('')).resolves.toBe(false)
+    expect(requiredIf('prop')('')).toBe(false)
   })
 
   it('should validate empty string when simple boolean condition not met', () => {
-    expect(requiredIf('')('')).resolves.toBe(true)
-  })
-
-  it('should return a promise when passed a promise condition', () => {
-    const response = requiredIf(promiseT)('')
-    expect(response).toHaveProperty('then') // is a promise
-  })
-
-  it('should validate value if condition is a truthy promise', async () => {
-    expect(await requiredIf(promiseT)('')).toBe(false)
-    expect(await requiredIf(promiseT)('someValue')).toBe(true)
-  })
-
-  it('should NOT validate value if condition is a falsy promise', async () => {
-    expect(await requiredIf(promiseF)('')).toBe(true)
+    expect(requiredIf('')('')).toBe(true)
   })
 
   it('should pass the value to the validation function', () => {
@@ -40,5 +24,22 @@ describe('requiredIf validator', () => {
     requiredIf(validator)('foo', 'bar')
     expect(validator).toHaveBeenCalledTimes(1)
     expect(validator).toHaveBeenCalledWith('foo', 'bar')
+  })
+
+  it('should have the correct `this` context', () => {
+    let that
+    const validator = jest.fn(function () { that = this })
+    const context = { foo: 'foo' }
+    requiredIf(validator).call(context, 'value', 'parentVM')
+    expect(that).toEqual(context)
+  })
+
+  it('should work with a ref', () => {
+    const prop = ref(false)
+    // make sure if passed a `false` ref, it returns `true` directly
+    expect(requiredIf(prop)(false)).toBe(true)
+    prop.value = true
+    expect(requiredIf(prop)('')).toBe(false)
+    expect(requiredIf(prop)('1')).toBe(true)
   })
 })