+++ packages/validators/src/raw/__tests__/requiredUnless.spec.js
@@ -1,15 +1,16 @@
 import requiredUnless from '../requiredUnless'
 import { T, F } from '../../../tests/fixtures'
+import { ref } from 'vue-demi'
 
 describe('requiredUnless validator', () => {
   it('should not validate if prop is falsy', () => {
-    expect(requiredUnless(F)('')).resolves.toBe(false)
-    expect(requiredUnless(F)('truthy value')).resolves.toBe(true)
+    expect(requiredUnless(F)('')).toBe(false)
+    expect(requiredUnless(F)('truthy value')).toBe(true)
   })
 
-  it('should not validate when prop condition is truthy', () => {
-    expect(requiredUnless(T)('')).resolves.toBe(true)
-    expect(requiredUnless(T)('truthy value')).resolves.toBe(true)
+  it('should not validate when prop condition is truthy', async () => {
+    expect(requiredUnless(T)('')).toBe(true)
+    expect(requiredUnless(T)('truthy value')).toBe(true)
   })
 
   it('should pass the value to the validation function', () => {
@@ -17,5 +18,21 @@ describe('requiredUnless validator', () => {
     requiredUnless(validator)('foo', 'bar')
     expect(validator).toHaveBeenCalledTimes(1)
     expect(validator).toHaveBeenCalledWith('foo', 'bar')
+  })
+
+  it('should have the correct `this` context', () => {
+    let that
+    const validator = jest.fn(function () { that = this })
+    const context = { foo: 'foo' }
+    requiredUnless(validator).call(context, '', '')
+    expect(that).toEqual(context)
+  })
+
+  it('should work with a ref', () => {
+    const prop = ref(true)
+    expect(requiredUnless(prop)(true)).toBe(true)
+    prop.value = false
+    expect(requiredUnless(prop)('')).toBe(false)
+    expect(requiredUnless(prop)('1')).toBe(true)
   })
 })