+++ packages/validators/src/raw/requiredUnless.js
@@ -1,17 +1,18 @@
 import { req } from './core'
+import { unwrap } from '../common'
 
 const validate = (prop, val) => !prop ? req(val) : true
 /**
  * Returns required if the passed property is falsy.
- * @param {Boolean | String | function(any): (Boolean | Promise<boolean>)} propOrFunction
- * @return {function(*): (Boolean | Promise<Boolean>)}
+ * @param {Boolean | String | function(any): Boolean | Ref<string | boolean>} propOrFunction
+ * @return {function(value: *, parentVM: object): Boolean}
  */
 export default function requiredUnless (propOrFunction) {
-  return async function requiredUnlessInternal (value, parentVM) {
+  return function (value, parentVM) {
     if (typeof propOrFunction !== 'function') {
-      return validate(propOrFunction, value)
+      return validate(unwrap(propOrFunction), value)
     }
-    const result = await propOrFunction.call(this, value, parentVM)
+    const result = propOrFunction.call(this, value, parentVM)
     return validate(result, value)
   }
 }