+++ packages/validators/src/withMessages/alphaNum.js
@@ -6,5 +6,8 @@ import alphaNum from '../raw/alphaNum'
  */
 export default {
   $validator: alphaNum,
-  $message: 'The value must be alpha-numeric'
+  $message: 'The value must be alpha-numeric',
+  $params: {
+    type: 'alphaNum'
+  }
 }