+++ packages/validators/src/withMessages/minLength.js
@@ -9,6 +9,6 @@ export default function (min) {
   return {
     $validator: minLength(min),
     $message: ({ $params }) => `This field should be at least ${$params.min} long`,
-    $params: { min }
+    $params: { min, type: 'minLength' }
   }
 }