+++ packages/validators/src/withMessages/url.js
@@ -6,5 +6,8 @@ import url from '../raw/url'
  */
 export default {
   $validator: url,
-  $message: 'The value is not a valid URL address'
+  $message: 'The value is not a valid URL address',
+  $params: {
+    type: 'url'
+  }
 }